fix(address): make the address in the list and detail page can be copied

+++ wallet/CBWallet/Controllers/Address/AddressListViewController.m
@@ -182,7 +182,7 @@
     
     // create address
     NSUInteger idx = self.addressStore.countAllAddresses;
-    DLog(@"all addresses count: %lu", idx);
+    DLog(@"all addresses count: %lu", (unsigned long)idx);
     NSString *addressString = [CBWAddress addressStringWithIdx:idx acountIdx:self.account.idx];
     [self p_saveAddressString:addressString withIdx:idx];
 }
